requirements.txt
numpy==2.4.3
opencv-python-headless==4.13.0.92
Pillow==12.1.1
torch==2.10.0
torchvision==0.25.0
ultralytics==8.4.21
websockets==16.0
